fix get_train/eval_dataloader batch_size is None
